Safari-611.1.10/LayoutTests/fast/viewport/viewport-legacy-xhtmlmp-remove-and-add-expected.txt

ALERT: viewport size 320x480 scale 1 with limits [1, 5] and userScalable true
ALERT: viewport size 320x480 scale 1 with limits [1, 5] and userScalable true
ALERT: viewport size 320x480 scale 1 with limits [1, 5] and userScalable true
